requirements.txt
openai==1.54.3
anthropic==0.39.0
rich==13.9.4
python-dotenv==1.0.1
